flask/requirements.txt
Flask==2.1.0
psycopg2-binary